Add files via upload
The COVID-19 pandemic takes variable shapes and forms, in terms of cases and deaths, in different regions and countries of the world. This analysis demonstrates that regions and countries with high human development index have higher cases and deaths per million population due to COVID-19. This is due to international connectedness and mobility of their population related to trade and tourism, and their vulnerability related to older populations and higher rates of non-communicable diseases. We have also identified that the burden of the pandemic is also variable among high- and middle-income countries due to differences in the governance of the pandemic, fragmentation of health systems, and socio-economic inequities.